requirements.txt
appdirs==1.4.4
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.3.0
pystray==0.19.5
pywin32==311
PyYAML==6.0.2
schedule==1.2.2
setuptools==80.9.0
six==1.17.0
Werkzeug==3.1.3